requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
google-generativeai
gTTS